Add code for terminating JITServer
Added new API destroyJITServer to shutdown the JITServer.
Also added code to gracefully terminate the Listener thread.
