+++ runtime/compiler/runtime/Listener.hpp
@@ -25,6 +25,7 @@
 
 #include "j9.h"
 #include "infra/Monitor.hpp"  // TR::Monitor
+#include "net/ServerStream.hpp"
 
  /**
     @class TR_Listener
@@ -34,15 +35,36 @@
     Typical sequence executed by a JITServer is:
     (1) Create a TR_Listener object with "allocate()" function
     (2) Start a listener thread with  listener->startListenerThread(javaVM);
- 
-    The current implementation does not provide code for nicely terminating the listener thread.
  */
+
+#define OPENJ9_LISTENER_POLL_TIMEOUT 100 // in milliseconds
+
+class BaseCompileDispatcher;
+
 class TR_Listener
    {
 public:
    TR_Listener();
    static TR_Listener* allocate();
    void startListenerThread(J9JavaVM *javaVM);
+   void stop();
+   /**
+      @brief Function called to deal with incoming connection requests
+
+      This function opens a socket (non-blocking), binds it and then waits for incoming
+      connection by polling on it with a timeout (see OPENJ9_LISTENER_POLL_TIMEOUT).
+      If it ever comes out of polling (due to timeout or a new connection request),
+      it checks the exit flag. If the flag is set, then the thread exits.
+      Otherwise, it establishes the connection using accept().
+      Once a connection is accepted a ServerStream object is created (receiving the newly
+      opened socket descriptor as a parameter) and passed to the compilation handler.
+      Typically, the compilation handler places the ServerStream object in a queue and
+      returns immediately so that other connection requests can be accepted.
+      Note: it must be executed on a separate thread as it needs to keep listening for new connections.
+
+      @param [in] compiler Object that defines the behavior when a new connection is accepted
+   */
+   void serveRemoteCompilationRequests(BaseCompileDispatcher *compiler);
    int32_t waitForListenerThreadExit(J9JavaVM *javaVM);
    void setAttachAttempted(bool b) { _listenerThreadAttachAttempted = b; }
    bool getAttachAttempted() const { return _listenerThreadAttachAttempted; }
@@ -61,6 +83,20 @@ private:
    j9thread_t _listenerOSThread;
    volatile bool _listenerThreadAttachAttempted;
    volatile bool _listenerThreadExitFlag;
+   };
+
+/**
+   @class BaseCompileDispatcher
+   @brief Abstract class defining the interface for the compilation handler
+
+   Typically, an user would derive this class and provide an implementation for "compile()"
+   An instance of the derived class needs to be passed to serveRemoteCompilationRequests
+   which internally calls "compile()"
+ */
+class BaseCompileDispatcher
+   {
+public:
+   virtual void compile(JITServer::ServerStream *stream) = 0;
    };
 
 #endif
