+++ runtime/oti/jitserver_error.h
@@ -30,4 +30,5 @@
 #define JITSERVER_STARTUP_FAILED -3
 #define JITSERVER_THREAD_ATTACH_FAILED -4
 #define JITSERVER_WAIT_TERM_FAILED -5
+#define JITSERVER_DESTROY_ERROR -6
 #endif /* jitserver_error_h */
